refactor: merge packages/shared into packages/server
After merging the bot package into server, the shared package only held
types, utilities, DB schema, logger, and the API service layer. These
are all server-side concerns that both server and web need access to,
so merging them into server simplifies the package graph from 3 packages
to 2.

Changes:
- Move all shared/ contents to server/src/shared/
- Server now exports ./shared for types, DB, logger, and API service
- Web imports from @alfira-bot/server/shared instead of @alfira-bot/shared
- Remove packages/shared entirely
- Update package.json scripts, Dockerfile, CI, drizzle config, and CLAUDE.md
