+++ .github/workflows/typecheck.yml
@@ -18,9 +18,6 @@ jobs:
       - name: Install dependencies
         run: bun install --frozen-lockfile
 
-      - name: Build shared package
-        run: bun run --filter @alfira-bot/shared build
-
       - name: Build server package
         run: bun run --filter @alfira-bot/server build
