+++ package.json
@@ -6,11 +6,11 @@
     "packages/*"
   ],
   "scripts": {
-    "dev": "bun run --filter @alfira-bot/shared build && bun run --filter @alfira-bot/server build && docker compose up --build",
+    "dev": "bun run --filter @alfira-bot/server build && docker compose up --build",
     "web:build": "bun run --filter @alfira-bot/web build",
-    "db:migrate": "bun run --filter @alfira-bot/shared db:migrate",
-    "db:generate": "bun run --filter @alfira-bot/shared db:generate",
-    "db:studio": "bun run --filter @alfira-bot/shared db:studio",
+    "db:migrate": "bun run --filter @alfira-bot/server db:migrate",
+    "db:generate": "bun run --filter @alfira-bot/server db:generate",
+    "db:studio": "bun run --filter @alfira-bot/server db:studio",
     "lint": "biome lint .",
     "lint:fix": "biome lint --write .",
     "format": "biome format --write .",
