+++ packages/server/src/lib/serialization.ts
@@ -1,4 +1,4 @@
-import type { Song } from '@alfira-bot/shared';
+import type { Song } from '../shared';
 
 // Accept both Date and string createdAt — Drizzle uses Date at the DB level,
 // but we serialize to ISO string for JSON serialization.
