+++ packages/server/tsconfig.json
@@ -6,10 +6,7 @@
     "ignoreDeprecations": "6.0",
     "baseUrl": ".",
     "rootDir": "./src",
-    "paths": {
-      "@alfira-bot/shared": ["./node_modules/@alfira-bot/shared/dist/"],
-      "@alfira-bot/shared/*": ["./node_modules/@alfira-bot/shared/dist/*"]
-    },
+    "paths": {},
     "lib": ["ES2022"],
     "types": ["bun-types", "node"],
     "outDir": "./dist",
