--- packages/shared/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@alfira-bot/shared",
-  "version": "0.1.0",
-  "private": true,
-  "main": "dist/index.js",
-  "exports": {
-    ".": "./dist/index.js",
-    "./api": "./dist/api.js",
-    "./format": "./dist/format.js",
-    "./logger": "./dist/logger.js",
-    "./db": "./dist/db/index.js"
-  },
-  "scripts": {
-    "build": "tsc && cp -r src/db/migrations dist/db/",
-    "db:generate": "cd ../.. && drizzle-kit generate",
-    "db:migrate": "cd ../.. && drizzle-kit migrate",
-    "db:migrate:deploy": "cd ../.. && drizzle-kit migrate",
-    "db:studio": "cd ../.. && drizzle-kit studio"
-  },
-  "dependencies": {
-    "drizzle-orm": "^0.45.2",
-    "pino": "^10.3.1"
-  },
-  "devDependencies": {
-    "bun-types": "^1.3.11",
-    "drizzle-kit": "^0.31.10",
-    "typescript": "^6.0.0"
-  }
-}
