+++ packages/web/src/components/AddSongsModal.tsx
@@ -1,5 +1,5 @@
-import type { PlaylistDetail, Song } from '@alfira-bot/shared';
-import { formatDuration } from '@alfira-bot/shared';
+import type { PlaylistDetail, Song } from '@alfira-bot/server/shared';
+import { formatDuration } from '@alfira-bot/server/shared';
 import { useVirtualizer } from '@tanstack/react-virtual';
 import { memo, useCallback, useEffect, useRef, useState } from 'react';
 import { addSongToPlaylist, getSongsPage } from '../api/api';
