+++ packages/web/src/components/SongCard.tsx
@@ -1,5 +1,5 @@
-import type { Playlist, Song } from '@alfira-bot/shared';
-import { formatDuration } from '@alfira-bot/shared';
+import type { Playlist, Song } from '@alfira-bot/server/shared';
+import { formatDuration } from '@alfira-bot/server/shared';
 import { CircleNotchIcon, HeadphonesIcon, PlayIcon } from '@phosphor-icons/react';
 import React, { useCallback, useMemo } from 'react';
 import { useSongEdit } from '../context/SongEditContext';
