+++ packages/web/src/hooks/useProgressBar.ts
@@ -1,4 +1,4 @@
-import type { QueueState } from '@alfira-bot/shared';
+import type { QueueState } from '@alfira-bot/server/shared';
 import { useCallback, useEffect, useRef, useState } from 'react';
 
 /**
